backup/backend/requirements.txt
# backend/requirements.txt
# Core dependencies
fastapi>=0.103.1
pydantic>=2.4.2
uvicorn>=0.23.2
python-multipart>=0.0.6
python-dotenv>=1.0.0
httpx>=0.25.0

# AI & ML
google-generativeai>=0.3.1
numpy>=1.24.0
langchain>=0.1.0
langchain-google-genai>=0.0.5
langchain-community>=0.0.10
llama-index>=0.9.5
sentence-transformers>=2.2.2
chromadb>=0.4.18

# Document processing
PyPDF2>=3.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
unstructured>=0.10.30

# Utilities
python-jose>=3.3.0
passlib>=1.7.4
Pillow>=10.0.0

# Testing
pytest>=7.4.2
pytest-asyncio>=0.21.1